+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-fewsjdbc
 2.26 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Let datasource.py return a DataFrame-based timeseries.
 
 
 2.25 (2013-04-10)